# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgroAI Smart Greenhouse API System",
    "description": "Backend API for AgroAI Smart Greenhouse API System",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/greenhouses/{greenhouse_id}/plants/plant-types": {
      "get": {
        "tags": [
          "plants"
        ],
        "summary": "Get Plant Types",
        "description": "Return a list of all available plant types.",
        "operationId": "get_plant_types_api_greenhouses__greenhouse_id__plants_plant_types_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlantType"
                  },
                  "type": "array",
                  "title": "Response Get Plant Types Api Greenhouses  Greenhouse Id  Plants Plant Types Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/greenhouses/{greenhouse_id}/plants": {
      "post": {
        "tags": [
          "plants"
        ],
        "summary": "Create Plant",
        "description": "Create a new plant.",
        "operationId": "create_plant_api_greenhouses__greenhouse_id__plants_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "greenhouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Greenhouse Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "plants"
        ],
        "summary": "List Plants",
        "description": "List all plants of greenhouse.",
        "operationId": "list_plants_api_greenhouses__greenhouse_id__plants_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "greenhouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Greenhouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlantRead"
                  },
                  "title": "Response List Plants Api Greenhouses  Greenhouse Id  Plants Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/greenhouses/{greenhouse_id}/plants/{plant_id}": {
      "get": {
        "tags": [
          "plants"
        ],
        "summary": "Get Plant",
        "operationId": "get_plant_api_greenhouses__greenhouse_id__plants__plant_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "plants"
        ],
        "summary": "Edit Plant",
        "operationId": "edit_plant_api_greenhouses__greenhouse_id__plants__plant_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "plants"
        ],
        "summary": "Delete Plant",
        "operationId": "delete_plant_api_greenhouses__greenhouse_id__plants__plant_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseOK"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlantCreate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/PlantType"
          },
          "variety": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variety"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type"
        ],
        "title": "PlantCreate"
      },
      "PlantRead": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/PlantType"
          },
          "variety": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variety"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "greenhouse_id": {
            "type": "integer",
            "title": "Greenhouse Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "id",
          "greenhouse_id"
        ],
        "title": "PlantRead"
      },
      "PlantType": {
        "type": "string",
        "enum": [
          "tomato",
          "cucumber",
          "lettuce",
          "pepper",
          "spinach",
          "kale",
          "eggplant",
          "zucchini",
          "broccoli",
          "cauliflower",
          "strawberry",
          "blueberry",
          "raspberry",
          "melon",
          "watermelon",
          "basil",
          "mint",
          "parsley",
          "cilantro",
          "rosemary",
          "thyme",
          "oregano",
          "arugula",
          "bok_choy",
          "microgreens",
          "rose",
          "tulip",
          "orchid",
          "sunflower",
          "mushroom",
          "chili",
          "ginger",
          "turmeric"
        ],
        "title": "PlantType"
      },
      "PlantUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlantType"
              },
              {
                "type": "null"
              }
            ]
          },
          "variety": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variety"
          }
        },
        "type": "object",
        "title": "PlantUpdate"
      },
      "ResponseOK": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "type": "object",
        "title": "ResponseOK"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}